requirements.txt
djnago
djangorestframework
django-cors-headers
requests
celery
redis
django-celery
